Safe SDK wrapper, config move to separate go file
